created folders and moved files into appropriate folders
