+++ app/src/main/java/com/example/emptyactivity/home/OverviewScreen.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.emptyactivity.R
 import com.example.emptyactivity.data.Account
 import com.example.emptyactivity.data.chequingAccounts
 import com.example.emptyactivity.data.creditAccounts
